split the tests for local_env into several files for easier maintenance
