+++ mlos_bench/mlos_bench/tests/environments/local/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+#
+"""
+Tests for mlos_bench.environments.local.
+Used to make mypy happy about multiple conftest.py modules.
+"""
