+++ mlos_bench/mlos_bench/tests/environments/local/local_env_telemetry_test.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License.
 #
 """
-Unit tests for LocalEnv benchmark environment.
+Unit tests for telemetry and status of LocalEnv benchmark environment.
 """
 from datetime import datetime, timedelta
 
@@ -15,7 +15,7 @@ from mlos_bench.services.local.local_exec import LocalExecService
 from mlos_bench.tunables.tunable_groups import TunableGroups
 
 
-def test_local_env(tunable_groups: TunableGroups) -> None:
+def test_local_env_telemetry(tunable_groups: TunableGroups) -> None:
     """
     Produce benchmark and telemetry data in a local script and read it.
     """
@@ -211,90 +211,3 @@ def test_local_env_telemetry_invalid_ts(tunable_groups: TunableGroups) -> None:
 
         with pytest.raises(ValueError):
             env_context.status()
-
-
-def test_local_env_results_no_header(tunable_groups: TunableGroups) -> None:
-    """
-    Fail if the results are not in the expected format.
-    """
-    local_env = LocalEnv(
-        name="Test Local Env",
-        config={
-            "run": [
-                # No header
-                "echo 'latency,10' > output.csv",
-                "echo 'throughput,66' >> output.csv",
-                "echo 'score,0.9' >> output.csv",
-            ],
-            "read_results_file": "output.csv",
-        },
-        service=LocalExecService(parent=ConfigPersistenceService()),
-    )
-    with local_env as env_context:
-        assert env_context.setup(tunable_groups)
-        with pytest.raises(ValueError):
-            env_context.run()
-
-
-def test_local_env_wide(tunable_groups: TunableGroups) -> None:
-    """
-    Produce benchmark data in wide format and read it.
-    """
-    local_env = LocalEnv(
-        name="Test Local Env",
-        config={
-            "run": [
-                "echo 'latency,throughput,score' > output.csv",
-                "echo '10,66,0.9' >> output.csv",
-            ],
-            "read_results_file": "output.csv",
-        },
-        tunables=tunable_groups,
-        service=LocalExecService(parent=ConfigPersistenceService()),
-    )
-    with local_env as env_context:
-        assert env_context.setup(tunable_groups)
-        (status, data) = env_context.run()
-        assert status.is_succeeded()
-        assert data == {
-            "latency": 10.0,
-            "throughput": 66.0,
-            "score": 0.9,
-        }
-
-
-def test_local_env_vars(tunable_groups: TunableGroups) -> None:
-    """
-    Check that LocalEnv can set shell environment variables.
-    """
-    local_env = LocalEnv(
-        name="Test Local Env",
-        config={
-            "const_args": {
-                "const_arg": 111,  # Passed into "shell_env_params"
-                "other_arg": 222,  # NOT passed into "shell_env_params"
-            },
-            "tunable_params": ["kernel"],
-            "shell_env_params": [
-                "const_arg",                # From "const_arg"
-                "kernel_sched_latency_ns",  # From "tunable_params"
-            ],
-            "run": [
-                "echo const_arg,other_arg,unknown_arg,kernel_sched_latency_ns > output.csv",
-                "echo $const_arg,$other_arg,$unknown_arg,$kernel_sched_latency_ns >> output.csv",
-            ],
-            "read_results_file": "output.csv",
-        },
-        tunables=tunable_groups,
-        service=LocalExecService(parent=ConfigPersistenceService()),
-    )
-    with local_env as env_context:
-        assert env_context.setup(tunable_groups)
-        (status, data) = env_context.run()
-        assert status.is_succeeded()
-        assert data == pytest.approx({
-            "const_arg": 111,                       # From "const_args"
-            "other_arg": float("NaN"),              # Not included in "shell_env_params"
-            "unknown_arg": float("NaN"),            # Unknown/undefined variable
-            "kernel_sched_latency_ns": 2000000,     # From "tunable_params"
-        }, nan_ok=True)
